# frontend/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KYC API",
    "version": "1.0.0",
    "description": "JSON:API endpoints for Applicant accounts and sessions, KYC applications and document evidence, reviewer workload, and access checks. Authenticated endpoints use the first-party session cookie. Error responses use a JSON:API errors array with safe, client-actionable details."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current frontend origin; /api/v1 is proxied to the Java API in development."
    }
  ],
...
  "paths": {
    "/api/v1/applicant-accounts": {
      "post": {
        "operationId": "createApplicantAccount",
        "summary": "Register an Applicant account",
        "requestBody": {
          "required": true,
          "content": {
            "application/vnd.api+json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicantAccountCreateDocument"
              }
            }
          },
          "description": "JSON:API credentials document. The email must be valid; the password must be at least six characters and include an uppercase letter and a special character."
        },
        "responses": {
          "201": {
            "description": "Account created. No session is established. Location identifies the new account resource.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                },
                "description": "URI of the newly created account resource."
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/CredentialInputInvalid"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "409": {
            "$ref": "#/components/responses/AccountConflict"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedRequestMediaType"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "tags": ["Account access"],
        "description": "Creates an account without signing the Applicant in. The request must use the applicant-accounts resource type. Email and password validation errors can return multiple entries with source.pointer values. The password must contain at least six characters, one uppercase letter, and one special character."
      }
    },
    "/api/v1/applicant-sessions": {
      "post": {
        "operationId": "createApplicantSession",
        "summary": "Authenticate an Applicant and establish a session",
        "requestBody": {
          "required": true,
          "content": {
            "application/vnd.api+json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicantSessionCreateDocument"
              }
            }
          },
          "description": "JSON:API credentials document for an Applicant account."
        },
        "responses": {
          "201": {
            "description": "Session established. Sets an HttpOnly session cookie and a JavaScript-readable XSRF-TOKEN cookie; returns the CSRF token in X-CSRF-TOKEN.",
            "headers": {
              "Location": {
                "description": "Canonical URL using the public derived session UUID, never the credential.",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              },
              "Set-Cookie": {
                "description": "Sets the opaque HttpOnly session cookie and the JavaScript-readable XSRF-TOKEN cookie.",
                "schema": {
                  "type": "string"
                }
              },
              "X-CSRF-TOKEN": {
                "description": "CSRF token also stored in the XSRF-TOKEN cookie. Send it as X-XSRF-TOKEN on authenticated state-changing requests.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantSessionDocument"
                },
                "description": "JSON:API session resource with public session UUID and journey URL. The session secret is only sent in the HttpOnly cookie."
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/CredentialInputInvalid"
          },
          "401": {
            "$ref": "#/components/responses/InvalidCredentials"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedRequestMediaType"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "tags": ["Account access"],
        "description": "Authenticates credentials and rotates any existing session. Invalid credentials share one 401 response so the API does not reveal whether an email address is registered. Success returns an opaque session resource, sets the session and XSRF-TOKEN cookies, and returns the CSRF token in X-CSRF-TOKEN."
      }
    },
    "/api/v1/applicant-sessions/current": {
      "delete": {
        "operationId": "deleteCurrentApplicantSession",
        "summary": "Revoke the current Applicant session",
        "responses": {
          "204": {
            "description": "Session revoked and session cookie cleared. No response body.",
            "headers": {
              "Set-Cookie": {
                "description": "Expires the current session cookie.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          }
        },
        "tags": ["Account access"],
        "description": "Revokes the session identified by the session cookie and clears that cookie. When a session cookie is present, send the matching CSRF token in X-XSRF-TOKEN.",
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-KYCSESSION",
        "description": "Opaque, secure, HttpOnly cookie set after successful sign-in. The browser sends it automatically to the API origin."
      }
    },
    "schemas": {
      "ApplicantAccountCreateDocument": {
        "type": "object",
        "required": ["data"],
        "properties": {
          "data": {
            "type": "object",
            "required": ["type", "attributes"],
            "properties": {
              "type": {
                "const": "applicant-accounts",
                "description": "Must be applicant-accounts."
              },
              "attributes": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Applicant email address."
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6,
                    "writeOnly": true,
                    "description": "Password is write-only. Use at least 6 characters including an uppercase letter and a special character."
                  }
                }
              }
            },
            "description": "Applicant account resource to create."
          }
        },
        "description": "JSON:API request document for Applicant registration."
      },
      "ApplicantSessionCreateDocument": {
        "type": "object",
        "required": ["data"],
        "properties": {
          "data": {
            "type": "object",
            "required": ["type", "attributes"],
            "properties": {
              "type": {
                "const": "applicant-sessions",
                "description": "Must be applicant-sessions."
              },
              "attributes": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Applicant email address."
                  },
                  "password": {
                    "type": "string",
                    "writeOnly": true,
                    "description": "Password is write-only. Supply the password used when registering the account.",
                    "minLength": 6
                  }
                }
              }
            },
            "description": "Applicant session resource to create."
          }
        },
        "description": "JSON:API request document for Applicant sign-in."
      },
      "ApplicantSessionDocument": {
        "type": "object",
        "required": ["data"],
        "properties": {
          "data": {
            "type": "object",
            "required": ["type", "id", "attributes"],
            "properties": {
              "type": {
                "const": "applicant-sessions"
              },
              "id": {
                "type": "string",
                "format": "uuid",
                "description": "Non-reversible public identifier derived from the session secret; not a credential."
              },
              "attributes": {
                "type": "object",
                "required": ["href"],
                "properties": {
                  "href": {
                    "type": "string",
                    "format": "uri-reference"
                  }
                }
              }
            }
          }
        },
        "description": "JSON:API response containing the public session identifier and Applicant journey URL. Authentication secret is stored in the session cookie."
      },
...
    },
    "responses": {
      "CredentialInputInvalid": {
        "description": "The credential document is malformed or contains invalid fields. The errors array can contain multiple field errors; source.pointer identifies each invalid JSON:API member.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "malformedDocument": {
                "summary": "Malformed JSON:API document",
                "value": {
                  "errors": [
                    {
                      "status": "400",
                      "code": "malformed-document",
                      "title": "Malformed JSON:API document",
                      "detail": "The request body is not a valid JSON:API document."
                    }
                  ]
                }
              },
              "invalidFields": {
                "summary": "Credential field validation errors",
                "value": {
                  "errors": [
                    {
                      "status": "400",
                      "code": "validation-failed",
                      "title": "Invalid attribute",
                      "detail": "Enter a valid email address.",
                      "source": {
                        "pointer": "/data/attributes/email"
                      }
                    },
                    {
                      "status": "400",
                      "code": "validation-failed",
                      "title": "Invalid attribute",
                      "detail": "Use at least 6 characters including an uppercase letter and special character.",
                      "source": {
                        "pointer": "/data/attributes/password"
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "AccountConflict": {
        "description": "An account already exists for the submitted email address. The response does not include account details.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "emailAlreadyRegistered": {
                "summary": "Email address already registered",
                "value": {
                  "errors": [
                    {
                      "status": "409",
                      "code": "email-already-registered",
                      "title": "Account conflict",
                      "detail": "An account already exists for this email address."
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "InvalidCredentials": {
        "description": "The email address or password is invalid. This response is the same for unknown email addresses and incorrect passwords.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "invalidCredentials": {
                "summary": "Credentials rejected",
                "value": {
                  "errors": [
                    {
                      "status": "401",
                      "code": "invalid-credentials",
                      "title": "Authentication failed",
                      "detail": "Invalid email or password"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "AuthenticationRequired": {
        "description": "No active Applicant session is available. Sign in and send the session cookie.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "authenticationRequired": {
                "summary": "Session required",
                "value": {
                  "errors": [
                    {
                      "status": "401",
                      "code": "authentication-required",
                      "title": "Authentication required",
                      "detail": "Sign in to continue."
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "AccessDenied": {
        "description": "The request is not authorized for this resource or is missing a valid CSRF token. The response does not disclose protected resource details.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "accessDenied": {
                "summary": "Access denied",
                "value": {
                  "errors": [
                    {
                      "status": "403",
                      "code": "access-denied",
                      "title": "Access denied",
                      "detail": "You do not have permission to access this resource."
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "RateLimited": {
        "description": "Too many registration or sign-in attempts were received. Retry after the number of seconds in the Retry-After header.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "rateLimited": {
                "summary": "Request rate limit reached",
                "value": {
                  "errors": [
                    {
                      "status": "429",
                      "code": "rate-limited",
                      "title": "Too many requests",
                      "detail": "Too many attempts. Try again later."
                    }
                  ]
                }
              }
            }
          }
        },
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying; value is at least 1.",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        }
      },
      "UnsupportedRequestMediaType": {
        "description": "The request Content-Type is not accepted for this operation. JSON:API request bodies must use application/vnd.api+json. For document-evidence uploads, use multipart/form-data; image validation failures return 422.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "unsupportedMediaType": {
                "summary": "Unsupported request media type",
                "value": {
                  "errors": [
                    {
                      "status": "415",
                      "code": "unsupported-media-type",
                      "title": "Unsupported media type",
                      "detail": "Use application/vnd.api+json for JSON:API request bodies."
                    }
                  ]
                }
              },
              "mediaTypeParameters": {
                "summary": "Unsupported JSON:API media type parameters",
                "value": {
                  "errors": [
                    {
                      "status": "415",
                      "code": "unsupported-media-type",
                      "title": "Unsupported media type",
                      "detail": "Use application/vnd.api+json without media type parameters."
                    }
                  ]
                }
              }
            }
          }
        }
      },
...
      "NotAcceptable": {
        "description": "The requested response representation is not supported. Accept application/vnd.api+json.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "notAcceptable": {
                "summary": "Unsupported response representation",
                "value": {
                  "errors": [
                    {
                      "status": "406",
                      "code": "not-acceptable",
                      "title": "Not acceptable",
                      "detail": "Accept application/vnd.api+json for JSON:API responses."
                    }
                  ]
                }
              }
            }
          }
        }
      },
...
    },
...
  }
}